% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{benthic.params-class}
\alias{benthic.params-class}
\alias{benthic.params}
\title{Benthic species parameters}
\description{
An S4 object holding the physiological, detritus-feeding, reproduction and
fishing parameters for a benthic species. Usually created with
\code{\link[=Setup.Benthic]{Setup.Benthic()}}.
}
