% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{detritus.params-class}
\alias{detritus.params-class}
\alias{detritus.params}
\title{Detritus parameters}
\description{
An S4 object holding the parameters of the detritus pool. Usually created
with \code{\link[=Setup.Detritus]{Setup.Detritus()}}.
}
