% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{detritus.results-class}
\alias{detritus.results-class}
\alias{detritus.results}
\title{Detritus results}
\description{
An S4 object holding the detritus pool output (biomass and detrital
in/out fluxes) read back from a run. Usually created by \code{\link[=Read.In]{Read.In()}}.
}
