% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{grid.params-class}
\alias{grid.params-class}
\alias{grid.params}
\title{Grid discretisation parameters}
\description{
An S4 object describing the mass, time and (optional) spatial discretisation
of a run. Usually created with \code{\link[=Setup.Grid]{Setup.Grid()}}.
}
