% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{pelagic.params-class}
\alias{pelagic.params-class}
\alias{pelagic.params}
\title{Pelagic species parameters}
\description{
An S4 object holding the physiological, feeding-preference, reproduction and
fishing parameters for a pelagic predator species. Usually created with
\code{\link[=Setup.Pelagic]{Setup.Pelagic()}}.
}
