% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{pelagic.results-class}
\alias{pelagic.results-class}
\alias{pelagic.results}
\title{Pelagic results}
\description{
An S4 object holding the pelagic species output (densities, growth,
mortality, predation, fishing and biomass budgets) read back from a run.
Usually created by \code{\link[=Read.In]{Read.In()}}.
}
