% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{plankton.params-class}
\alias{plankton.params-class}
\alias{plankton.params}
\title{Plankton parameters}
\description{
An S4 object holding the size range and spectrum parameters for the plankton
resource. Usually created with \code{\link[=Setup.Plankton]{Setup.Plankton()}}.
}
