% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{plankton.results-class}
\alias{plankton.results-class}
\alias{plankton.results}
\title{Plankton results}
\description{
An S4 object holding the plankton size-spectrum output read back from a run.
Usually created by \code{\link[=Read.In]{Read.In()}}.
}
