% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{run.params-class}
\alias{run.params-class}
\alias{run.params}
\title{Run-level parameters}
\description{
An S4 object holding settings that apply to an entire DBPM run, such as the
number of pelagic and benthic species, the spatial dimension and the
integration method. Usually created with \code{\link[=Setup.Run]{Setup.Run()}}.
}
