% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.Detritus.r
\name{Setup.Detritus}
\alias{Setup.Detritus}
\title{Set up the detritus pool}
\usage{
Setup.Detritus(
  run.in,
  w_0 = 0.6,
  initial_flag = FALSE,
  ts_flag = FALSE,
  filename
)
}
\arguments{
\item{run.in}{A \link{run.params} object.}

\item{w_0}{Initial detritus biomass.}

\item{initial_flag, ts_flag}{Logical flags indicating whether
initial-condition and time-series input files are supplied.}

\item{filename}{Character string naming the detritus pool (and its output
directory). Required.}
}
\value{
A \link{detritus.params} object.
}
\description{
Builds a \link{detritus.params} object and creates the detritus output directory
under the run directory.
}
\seealso{
\code{\link[=Setup.Benthic]{Setup.Benthic()}}
}
