% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.Fishing.r
\name{Setup.fishing}
\alias{Setup.fishing}
\title{Write fishing input for a species}
\usage{
Setup.fishing(species, run.in, grid.in, func, dataname)
}
\arguments{
\item{species}{A pelagic or benthic parameter object with \code{fishing_flag} set.}

\item{run.in}{A \link{run.params} object.}

\item{grid.in}{A \link{grid.params} object.}

\item{func}{A function of \verb{(m, t, x, y)} returning fishing mortality.}

\item{dataname}{Path to a CSV file of fishing values (first line is the file
type).}
}
\value{
Invisibly \code{NULL}; called for the side effect of writing the input
file under the run's \code{Input} directory.
}
\description{
Generates the \verb{_fishing_ts.txt} fishing-mortality time-series input file for a
pelagic or benthic species whose \code{fishing_flag} is \code{TRUE}. The values may be
supplied as a function of mass/time/space or read from a CSV data file.
Exactly one of \code{func} or \code{dataname} must be given.
}
\seealso{
\code{\link[=Setup.ts]{Setup.ts()}}, \code{\link[=Setup.Rep]{Setup.Rep()}}
}
