% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.Grid.r
\name{Setup.Grid}
\alias{Setup.Grid}
\title{Set up the grid discretisation}
\usage{
Setup.Grid(
  run,
  mmin = -28,
  mmax = 14,
  mstep = 0.2,
  moutstep = 1,
  t1 = 0,
  tmax = 1,
  tstep = (1/365),
  toutmin = 0,
  toutmax = tmax,
  toutstep = (73/365),
  xmin = 0,
  xmax = 1000,
  xstep = 50,
  xoutstep = 50,
  ymin = 0,
  ymax = 1000,
  ystep = 50,
  youtstep = 50
)
}
\arguments{
\item{run}{A \link{run.params} object (used to decide the spatial dimension).}

\item{mmin, mmax, mstep, moutstep}{Minimum, maximum, step and output step of the
log-mass grid.}

\item{t1, tmax, tstep}{Start, end and step of the time grid (in years).}

\item{toutmin, toutmax, toutstep}{Minimum, maximum and step of the time output
grid (in years).}

\item{xmin, xmax, xstep, xoutstep}{Minimum, maximum, step and output step of the
x-space grid.}

\item{ymin, ymax, ystep, youtstep}{Minimum, maximum, step and output step of the
y-space grid.}
}
\value{
A \link{grid.params} object.
}
\description{
Builds a \link{grid.params} object describing the mass, time and spatial
discretisation of a run. Spatial dimensions not used by \code{run} are collapsed
to a single trivial cell.
}
\seealso{
\code{\link[=Setup.Run]{Setup.Run()}}
}
