% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.Rep.r
\name{Setup.Rep}
\alias{Setup.Rep}
\title{Write reproduction input for a species}
\usage{
Setup.Rep(species, run.in, grid.in, func, dataname)
}
\arguments{
\item{species}{A pelagic or benthic parameter object.}

\item{run.in}{A \link{run.params} object.}

\item{grid.in}{A \link{grid.params} object.}

\item{func}{A function of \verb{(t, x, y)} returning reproduction values.}

\item{dataname}{Path to a CSV file of reproduction values (first line is the
file type).}
}
\value{
Invisibly \code{NULL}; called for the side effect of writing the input
file under the run's \code{Input} directory.
}
\description{
Generates the \verb{_rep_ts.txt} reproduction time-series input file for a pelagic
or benthic species whose \code{rep_method} is \code{1}. The values may be supplied as a
function of time/space or read from a CSV data file. Exactly one of \code{func} or
\code{dataname} must be given.
}
\note{
This function is only needed for species with \code{rep_method = 1}, which
require a reproduction time series. The package does not ship these input
data; the user must generate them here (from \code{func} or \code{dataname}) before
calling \code{\link[=SizeSpectrum]{SizeSpectrum()}}.
}
\seealso{
\code{\link[=Setup.ts]{Setup.ts()}}, \code{\link[=Setup.fishing]{Setup.fishing()}}
}
