% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{singlefile.results-class}
\alias{singlefile.results-class}
\alias{singlefile.results}
\title{Single-file results}
\description{
An S4 object holding the contents of a single named output file read back
from a run. Created by \code{\link[=Read.In]{Read.In()}} when a \code{filename} is supplied.
}
