% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{timestep.data-class}
\alias{timestep.data-class}
\alias{timestep.data}
\title{Single time-step data}
\description{
An S4 object holding the size spectrum at a single time step (or averaged
over a time window). Created by \code{\link[=Extract.Time]{Extract.Time()}} and \code{\link[=Average.Time]{Average.Time()}} and
consumed by \code{\link[=Plot.Spectrum]{Plot.Spectrum()}} and \code{\link[=Points.Spectrum]{Points.Spectrum()}}.
}
